requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
requests
selenium
webdriver-manager
torch
transformers
pillow
faiss-cpu
python-dotenv
numpy